# docs/swagger.yaml
basePath: /
definitions:
  echo.HTTPError:
    properties:
      message: {}
    type: object
...
  github_com_max-main-team_backend_hackaton_MAX_internal_http_dto.DepartmentInfoResponse:
    properties:
      code:
        example: 09.03.02
        type: string
      department_name:
        example: Sowtware Engineering
        type: string
      faculty_name:
        example: FITIP
        type: string
      id:
        example: 123456789
        type: integer
    required:
    - code
    - department_name
    - faculty_name
    - id
    type: object
...
  github_com_max-main-team_backend_hackaton_MAX_internal_http_dto.FacultyInfoResponse:
    properties:
      id:
        type: integer
      name:
        type: string
      university_name:
        type: string
    type: object
  github_com_max-main-team_backend_hackaton_MAX_internal_http_dto.GroupInfoResponse:
    properties:
      code:
        type: string
      course_id:
        type: integer
      department_name:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
...
  github_com_max-main-team_backend_hackaton_MAX_internal_http_dto.User:
    properties:
      first_name:
        example: John
        type: string
      id:
        example: 123456789
        type: integer
      language_code:
        example: ru
        type: string
      last_name:
        example: Doe
        type: string
      photo_url:
        example: https://example.com/photo.jpg
        type: string
      username:
        example: johndoe
        type: string
    required:
    - first_name
    - id
    type: object
...
  github_com_max-main-team_backend_hackaton_MAX_internal_models_http_personalities.AcceptAccessRequest:
    properties:
      course_group_id:
        type: integer
      faculty_id:
        type: integer
      role:
        $ref: '#/definitions/github_com_max-main-team_backend_hackaton_MAX_internal_models_repository_personalities.RoleType'
      university_department_id:
        type: integer
      university_id:
        type: integer
      user_id:
        type: integer
    required:
    - role
    - user_id
    type: object
  github_com_max-main-team_backend_hackaton_MAX_internal_models_http_personalities.AccessRequestResponse:
    properties:
      data:
        items:
          properties:
            first_name:
              type: string
            last_name:
              type: string
            request_id:
              type: integer
            role:
              $ref: '#/definitions/github_com_max-main-team_backend_hackaton_MAX_internal_models_repository_personalities.RoleType'
            user_id:
              type: integer
            username:
              type: string
          type: object
        type: array
      has_more:
        type: boolean
    type: object
  github_com_max-main-team_backend_hackaton_MAX_internal_models_http_personalities.RequestAccessToUniversity:
    properties:
      role:
        $ref: '#/definitions/github_com_max-main-team_backend_hackaton_MAX_internal_models_repository_personalities.RoleType'
      university_id:
        type: integer
    type: object
...
info:
  contact: {}
  description: API for Hackaton backend
  title: Hackaton API
  version: "1.0"
paths:
...
  /admin/personalities/access:
    delete:
      consumes:
      - application/json
      description: Decline access request for user. Admin role required.
      parameters:
      - description: Request ID
        in: query
        name: request_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "401":
          description: Unauthorized user
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - BearerAuth: []
      summary: Reject access request
      tags:
      - personalities
    get:
      consumes:
      - application/json
      description: Get all access requests to join university (student/teacher/administration).
        Admin role required.
      parameters:
      - default: 5
        description: Limit of requests, max(50), default(5)
        in: query
        name: limit
        required: true
        type: integer
      - default: 0
        description: Offset, default(0)
        in: query
        name: offset
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Requests for administration
          schema:
            $ref: '#/definitions/github_com_max-main-team_backend_hackaton_MAX_internal_models_http_personalities.AccessRequestResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "401":
          description: Unauthorized user
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - BearerAuth: []
      summary: Get all access requests for administration
      tags:
      - personalities
    post:
      consumes:
      - application/json
      description: Current authenticated user sends a request to get a role in a university
        (student/teacher/administration).
      parameters:
      - description: Access request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_max-main-team_backend_hackaton_MAX_internal_models_http_personalities.RequestAccessToUniversity'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "401":
          description: Unauthorized user
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Request access to join a university
      tags:
      - personalities
  /admin/personalities/access/accept:
    post:
      consumes:
      - application/json
      description: Accept Request of user that want to be (student/teacher/administration),
        for student field university_department_id is required, course_group_id can
        be skipped. For administrations university_id is required, faculty_id can
        be skipped
      parameters:
      - description: Access request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_max-main-team_backend_hackaton_MAX_internal_models_http_personalities.AcceptAccessRequest'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "401":
          description: Unauthorized user
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Accept Request for adding in University
      tags:
      - personalities
...
  /personalities/departments:
    get:
      consumes:
      - application/json
      description: Get all departments for faculty by faculty ID
      parameters:
      - description: Faculty ID
        in: query
        name: faculty_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Departments
          schema:
            items:
              $ref: '#/definitions/github_com_max-main-team_backend_hackaton_MAX_internal_http_dto.DepartmentInfoResponse'
            type: array
        "400":
          description: Invalid request parameter
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "401":
          description: Unauthorized user
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Get all departments for faculty
      tags:
      - personalities
  /personalities/faculty:
    get:
      consumes:
      - application/json
      description: Get all faculties for university by university ID
      parameters:
      - description: University ID
        in: query
        name: university_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Faculties
          schema:
            items:
              $ref: '#/definitions/github_com_max-main-team_backend_hackaton_MAX_internal_http_dto.FacultyInfoResponse'
            type: array
        "400":
          description: Invalid request parameter
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "401":
          description: Unauthorized user
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Get all faculties for university
      tags:
      - personalities
  /personalities/groups:
    get:
      consumes:
      - application/json
      description: Get all course groups for department by department ID
      parameters:
      - description: Department ID
        in: query
        name: department_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Groups
          schema:
            items:
              $ref: '#/definitions/github_com_max-main-team_backend_hackaton_MAX_internal_http_dto.GroupInfoResponse'
            type: array
        "400":
          description: Invalid request parameter
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "401":
          description: Unauthorized user
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Get all groups for department
      tags:
      - personalities
  /personalities/student:
    get:
      consumes:
      - application/json
      description: Get all students for course group by group ID
      parameters:
      - description: Course Group ID
        in: query
        name: group_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Students
          schema:
            items:
              $ref: '#/definitions/github_com_max-main-team_backend_hackaton_MAX_internal_http_dto.User'
            type: array
        "400":
          description: Invalid request parameter
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "401":
          description: Unauthorized user
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Get all students for group
      tags:
      - personalities
  /personalities/teachers:
    get:
      consumes:
      - application/json
      description: Get all teachers for university by university ID
      parameters:
      - description: University ID
        in: query
        name: university_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Teachers
          schema:
            items:
              $ref: '#/definitions/github_com_max-main-team_backend_hackaton_MAX_internal_http_dto.User'
            type: array
        "400":
          description: Invalid request parameter
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "401":
          description: Unauthorized user
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Get all teachers for university
      tags:
      - personalities
  /personalities/universities:
    get:
      consumes:
      - application/json
      description: Get all universities where the authenticated user has access (as
        admin or student)
      produces:
      - application/json
      responses:
        "200":
          description: Universities
          schema:
            items:
              $ref: '#/definitions/github_com_max-main-team_backend_hackaton_MAX_internal_http_dto.UniInfoResponse'
            type: array
        "401":
          description: Unauthorized user
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Get all universities for authenticated person
      tags:
      - personalities
...
swagger: "2.0"